+++ hooks/app-add-v4-services.js
@@ -13,7 +13,7 @@ module.exports = async (app, lando) => {
 
   // if no service is set as the primary one lets set the first one as primary
   if (_.find(app.v4.parsedConfig, service => service.primary === true) === undefined) {
-    if (_.has(app, 'v4.parsedConfig[0.name')) app.v4.parsedConfig[0].primary = true;
+    if (_.has(app, 'v4.parsedConfig[0].name')) app.v4.parsedConfig[0].primary = true;
   }
 
   // note the primary service in a more convenient place so we dont have to search for it all the time
@@ -62,17 +62,7 @@ module.exports = async (app, lando) => {
     app.info.push(service.info);
   });
 
-  // emit an event so other plugins can augment the servies with additonal things before we get their data
-  return app.events.emit('pre-services-generate', app.v4.services).then(services => {
-    // handle top level volumes and networks here
-    if (!_.isEmpty(app.config.volumes)) app.v4.addVolumes(app.config.volumes);
-    if (!_.isEmpty(app.config.networks)) app.v4.addNetworks(app.config.networks);
-
-    // then generate the orchestrator files for each service
-    _.forEach(app.v4.services, service => {
-      app.add(service.generateOrchestorFiles());
-      // Log da things
-      app.log.debug('generated v4 %s service %s', service.type, service.name);
-    });
-  });
+  // handle top level volumes and networks here
+  if (!_.isEmpty(app.config.volumes)) app.v4.addVolumes(app.config.volumes);
+  if (!_.isEmpty(app.config.networks)) app.v4.addNetworks(app.config.networks);
 };