+++ index.js
@@ -3,7 +3,6 @@
 // Modules
 const _ = require('lodash');
 const fs = require('fs');
-const ip = require('ip');
 const path = require('path');
 
 // Default env values
@@ -45,11 +44,10 @@ module.exports = async lando => {
 
   // certs stuff
   // @TODO: should this end up elsewhere?
+  const caName = `${_.capitalize(lando.config.product)}CA`;
   const caDomain = lando.config.domain;
-  const caCert = path.join(caDir, `${caDomain}.pem`);
-  const caKey = path.join(caDir, `${caDomain}.key`);
-  const caProject = `landocasetupkenobi38ahsoka${lando.config.instance}`;
-  const certData = {caCert, caDir, caDomain, caKey, caProject};
+  const caCert = path.join(caDir, `${caName}.crt`);
+  const caKey = path.join(caDir, `${caName}.key`);
 
   // Ensure some dirs exist before we start
   _.forEach([binDir, caDir, sshDir], dir => fs.mkdirSync(dir, {recursive: true}));
@@ -57,20 +55,18 @@ module.exports = async lando => {
   // Ensure we munge plugin stuff together appropriately
   lando.events.once('pre-install-plugins', async options => await require('./hooks/lando-setup-common-plugins')(lando, options)); // eslint-disable-line max-len
 
-  // Ensure we setup docker-compose if needed
-  lando.events.once('pre-setup', async options => await require('./hooks/lando-setup-orchestrator')(lando, options)); // eslint-disable-line max-len
+  // move v3 scripts directories as needed
+  lando.events.on('pre-setup', 0, async () => await require('./hooks/lando-copy-v3-scripts')(lando));
 
   // Ensure we setup docker if needed
-  lando.events.once('pre-setup', async options => {
-    switch (process.platform) {
-      case 'darwin':
-        return await require('./hooks/lando-setup-build-engine-darwin')(lando, options);
-      case 'linux':
-        return await require('./hooks/lando-setup-build-engine-linux')(lando, options);
-      case 'win32':
-        return await require('./hooks/lando-setup-build-engine-win32')(lando, options);
-    }
-  });
+  lando.events.once('pre-setup', async options => await require(`./hooks/lando-setup-build-engine-${process.platform}`)(lando, options)); // eslint-disable-line max-len
+
+  // Ensure we create and install ca if needed
+  lando.events.once('pre-setup', async options => await require('./hooks/lando-setup-create-ca')(lando, options));
+  lando.events.once('pre-setup', async options => await require(`./hooks/lando-setup-install-ca-${process.platform}`)(lando, options)); // eslint-disable-line max-len
+
+  // Ensure we setup docker-compose if needed
+  lando.events.once('pre-setup', async options => await require('./hooks/lando-setup-orchestrator')(lando, options));
 
   // make sure Lando Specification 337 is available to all
   lando.events.on('post-bootstrap-app', async () => await require('./hooks/lando-add-l337-spec')(lando));
@@ -90,13 +86,8 @@ module.exports = async lando => {
   // autostart docker if we need to
   lando.events.once('engine-autostart', async () => await require('./hooks/lando-autostart-engine')(lando));
 
-  // Make sure we have a host-exposed root ca if we don't already
-  // NOTE: we don't run this on the caProject otherwise infinite loop happens!
-  lando.events.on('pre-engine-start', 2, async data => await require('./hooks/lando-setup-ca')(lando, data, certData));
-
-  // Let's also make a copy of caCert with the standarized .crt ending for better linux compat
-  // See: https://github.com/lando/lando/issues/1550
-  lando.events.on('pre-engine-start', 3, async () => await require('./hooks/lando-copy-ca')(lando, certData));
+  // move v3 scripts directories as needed
+  lando.events.on('pre-engine-start', 0, async () => await require('./hooks/lando-copy-v3-scripts')(lando));
 
   // Return some default things
   return _.merge({}, defaults, uc(), {config: {
@@ -107,7 +98,7 @@ module.exports = async lando => {
       LANDO_DOMAIN: lando.config.domain,
       LANDO_HOST_HOME: lando.config.home,
       LANDO_HOST_OS: lando.config.os.platform,
-      LANDO_HOST_IP: (process.platform === 'linux') ? ip.address() : 'host.docker.internal',
+      LANDO_HOST_IP: 'host.lando.internal',
       LANDO_LEIA: _.toInteger(lando.config.leia),
       LANDO_MOUNT: '/app',
     },
@@ -118,7 +109,6 @@ module.exports = async lando => {
     caCert,
     caDomain,
     caKey,
-    caProject,
     maxKeyWarning: 10,
   }});
 };