+++ plugins/proxy/app.js
@@ -36,6 +36,12 @@ const getAllPorts = (noHttp = false, noHttps = false, config) => {
   return _.flatten(ports).join(', ');
 };
 
+const hasCerts = (app, id) => {
+  const info = app.info.find(service => service.service === id);
+  const v4 = _.get(app, 'v4.services', []).find(service => service.id === id);
+  return info?.hasCerts === true || v4?.certs !== false;
+};
+
 /*
  * Helper to scanPorts
  */
@@ -91,9 +97,14 @@ module.exports = (app, lando) => {
       _.forEach(files, file => lando.yaml.dump(file.path, file.data));
     });
 
-
     // Start and setup the proxy and services
-    app.events.on('pre-start', 1, () => {
+    app.events.on('pre-start', 1, async () => {
+      // generate proxy cert
+      await lando.generateCert('proxy._lando_', {domains: [
+        `*.${lando.config.proxyDomain}`,
+        'proxy._lando_.internal',
+      ]});
+
       // Determine what ports we need to discover
       const protocolStatus = utils.needsProtocolScan(lando.config.proxyCurrentPorts, lando.config.proxyLastPorts);
       // And then discover!
@@ -123,7 +134,7 @@ module.exports = (app, lando) => {
         }
 
         // Get list of services that *should* have certs for SSL
-        const sslReady = _(_.get(app, 'config.services', {}))
+        const sslReady = _(_.get(app, 'config.services', []))
           .map((data, name) => _.merge({}, data, {name}))
           .filter(data => data.ssl)
           .map(data => data.name)
@@ -140,11 +151,17 @@ module.exports = (app, lando) => {
         // Add hasCerts to servedBys
         _.forEach(servedBy, name => {
           const service = _.find(app.info, {service: name});
-          service.hasCerts = true;
+          if (service) service.hasCerts = true;
         });
 
+        // get new v4 ssl ready services
+        const sslReadyV4 = _(_.get(app, 'v4.services', []))
+          .filter(data => data.certs)
+          .map(data => data.id)
+          .value();
+
         // Parse config
-        return utils.parseConfig(app.config.proxy, _.compact(_.flatten([sslReady, servedBy])));
+        return utils.parseConfig(app.config.proxy, _.compact(_.flatten([sslReady, servedBy, sslReadyV4])));
       })
 
       // Map to docker compose things
@@ -201,7 +218,7 @@ module.exports = (app, lando) => {
             .flatMap(s => s.urls = _.uniq(s.urls.concat(utils.parse2Info(
               app.config.proxy[s.service],
               ports,
-              _.get(s, 'hasCerts', false),
+              hasCerts(app, s.service),
             ))))
             .value();
         }