+++ tasks/plugin-logout.js
@@ -3,6 +3,7 @@
 module.exports = lando => {
   return {
     command: 'plugin-logout',
+    usage: '$0 plugin-logout',
     level: 'tasks',
     run: async options => {
       const write = require('../utils/write-file');